[arp]: Extend proxy_arp to cover dualtor scenario (#6465)
- Rename `test_arp_dualtor.py` to `test_arp_extended.py`
- Create new file `test_arp_dualtor.py` specifically for dual ToR scenarios
    - Add test to verify proxy ARP behavior for target IPs learned on standby links
    - Skip this test on non 202012 devices since it depends on `accept_untracked_na` which is currently only available in 202012
- Extend GARP service to send NA messages
- Edit PR tests/kvm tests to use the new `test_arp_extended.py` name
