+++ tests/scripts/garp_service.py
@@ -5,7 +5,7 @@ import ptf.testutils as testutils
 import time
 
 from ipaddress import ip_interface
-from scapy.all import conf
+from scapy.all import conf, Ether, IPv6, ICMPv6ND_NA, ICMPv6NDOptSrcLLAddr
 from scapy.arch import get_if_hwaddr
 
 class GarpService:
@@ -28,7 +28,11 @@ class GarpService:
             intf_name = 'eth{}'.format(port)
             source_mac = get_if_hwaddr(intf_name)
             source_ip_str = config['target_ip']
+            source_ipv6_str = config['target_ipv6']
+            dut_mac = config['dut_mac']
+            dst_ipv6 = config['dst_ipv6']
             source_ip = str(ip_interface(source_ip_str).ip)
+            source_ipv6 = str(ip_interface(source_ipv6_str).ip)
 
             # PTF uses Scapy to create packets, so this is ok to create
             # packets through PTF even though we are using Scapy to send the packets
@@ -37,7 +41,13 @@ class GarpService:
                                                 ip_snd=source_ip,
                                                 ip_tgt=source_ip, # Re-use server IP as target IP, since it is within the subnet of the VLAN IP
                                                 arp_op=2)
-            self.packets[intf_name] = garp_pkt
+
+            na_pkt = Ether(src=source_mac, dst=dut_mac) \
+                   / IPv6(dst=dst_ipv6, src=source_ipv6) \
+                   / ICMPv6ND_NA(tgt=source_ipv6, S=1, R=0, O=0) \
+                   / ICMPv6NDOptSrcLLAddr(type=2, lladdr=source_mac)
+
+            self.packets[intf_name] = [garp_pkt, na_pkt]
 
     def send_garp_packets(self):
         '''
@@ -54,8 +64,9 @@ class GarpService:
 
         try:
             while True:
-                for socket, packet in sockets.items():
-                    socket.send(packet)
+                for socket, packet_list in sockets.items():
+                    for packet in packet_list:
+                        socket.send(packet)
 
                 if self.interval is None:
                     break
